commands: move init code to init.rs
